Move bmc_util to goto_checker module
