+++ src/cbmc/all_properties.cpp
@@ -14,7 +14,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <algorithm>
 #include <chrono>
 
-#include <cbmc/bmc_util.h>
+#include <goto-checker/bmc_util.h>
 
 #include <util/xml.h>
 #include <util/json.h>
