+++ src/goto-checker/bmc_util.cpp
