+++ src/goto-checker/module_dependencies.txt
@@ -1,5 +1,7 @@
+cbmc # symex_bmc will be moved next
 goto-checker
 goto-programs
 goto-symex
+linking
 solvers
 util
